# demos/sample-rest-stockapp/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api Documentation",
        "version": "1.0",
        "title": "Api Documentation",
        "termsOfService": "urn:tos",
        "contact": {},
        "license": { "name": "Apache 2.0", "url": "http://www.apache.org/licenses/LICENSE-2.0" }
    },
    "host": "localhost:8080",
    "basePath": "/",
...
    "paths": {
...
        "/v1/batch": {
            "post": {
                "tags": ["Batch Read Operations"],
                "summary": "Return multiple records from the server in a single request.",
                "operationId": "performBatchGet",
                "consumes": ["application/json", "application/msgpack"],
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "batchKeys",
                        "description": "batchKeys",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/BatchReadRequest" }
                        }
                    },
                    {
                        "name": "consistencyLevel",
                        "in": "query",
                        "description": "How replicas should be consulted in a read operation to provide the desired consistency guarantee.",
                        "required": false,
                        "type": "string",
                        "default": "CONSISTENCY_ONE",
                        "enum": ["CONSISTENCY_ONE", "CONSISTENCY_ALL"]
                    },
                    {
                        "name": "linearizeRead",
                        "in": "query",
                        "description": "Force reads to be linearized for server namespaces that support strong consistency mode.",
                        "required": false,
                        "type": "boolean",
                        "default": false,
                        "x-example": false
                    },
                    {
                        "name": "replica",
                        "in": "query",
                        "description": "Replica algorithm used to determine the target node for a single record command.",
                        "required": false,
                        "type": "string",
                        "default": "SEQUENCE",
                        "enum": ["MASTER", "MASTER_PROLES", "SEQUENCE", "RANDOM"]
                    },
                    {
                        "name": "sendKey",
                        "in": "query",
                        "description": "Send user defined key in addition to hash digest on both reads and writes.",
                        "required": false,
                        "type": "boolean",
                        "default": false,
                        "x-example": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/BatchRead" }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters or request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "403": {
                        "description": "Not authorized to access the resource",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "Non existent namespace used in one or more key.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
...
        "/v1/info": {
            "post": {
                "tags": ["Info Operations"],
                "summary": "Send a list of info commands to a random node in the cluster",
                "operationId": "infoAny",
                "consumes": ["application/json", "application/msgpack"],
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "infoRequests",
                        "description": "An array of info commands to send to the server. See https://www.aerospike.com/docs/reference/info/ for a list of valid commands.",
                        "required": true,
                        "schema": { "type": "array", "items": { "type": "string" } },
                        "x-examples": { "application/json": "['build', 'edition']" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commands sent succesfully.",
                        "schema": { "type": "object", "additionalProperties": { "type": "string" } }
                    },
                    "403": {
                        "description": "Not authorized to perform the info command",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
        "/v1/info/{node}": {
            "post": {
                "tags": ["Info Operations"],
                "summary": "Send a list of info commands to a specific node in the cluster.",
                "operationId": "infoNode",
                "consumes": ["application/json", "application/msgpack"],
                "produces": ["application/json", "application/msgpack"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "infoRequests",
                        "description": "An array of info commands to send to the server. See https://www.aerospike.com/docs/reference/info/ for a list of valid commands.",
                        "required": true,
                        "schema": { "type": "array", "items": { "type": "string" } },
                        "x-examples": { "application/json": "[build, edition]" }
                    },
                    {
                        "name": "nodeName",
                        "in": "path",
                        "description": "The node ID for the node which will receive the info commands.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commands sent succesfully.",
                        "schema": { "type": "object", "additionalProperties": { "type": "string" } }
                    },
                    "403": {
                        "description": "Not authorized to perform the info command",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "404": {
                        "description": "The specified Node does not exist.",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    },
                    "500": {
                        "description": "The Rest Client encountered an error processing the request",
                        "schema": { "$ref": "#/definitions/RestClientError" }
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
        "BatchRead": {
            "type": "object",
            "properties": {
                "binNames": { "type": "array", "items": { "type": "string" } },
                "key": { "$ref": "#/definitions/Key" },
                "readAllBins": { "type": "boolean" },
                "record": { "$ref": "#/definitions/Record" }
            },
            "title": "BatchRead"
        },
        "BatchReadRequest": {
            "type": "object",
            "properties": {
                "binNames": {
                    "type": "array",
                    "description": "List of bins to limit the record response to.",
                    "items": { "type": "string" }
                },
                "key": { "description": "Key to retrieve a record", "$ref": "#/definitions/Key" },
                "readAllBins": {
                    "type": "boolean",
                    "description": "Whether all bins should be returned with this record"
                }
            },
            "title": "BatchReadRequest"
        },
        "Key": {
            "type": "object",
            "required": ["namespace"],
            "properties": {
                "digest": {
                    "type": "string",
                    "example": "AAAAAAAAAAAAAAAAAAAAAAAAAAA=",
                    "description": "URL safe base64 encoded key digest. Returned by the server on batch responses. May be provided by client."
                },
                "keytype": {
                    "type": "string",
                    "example": "STRING",
                    "description": "Enum describing the type of the userKey. This field is omitted in MessagePack responses.",
                    "enum": ["STRING", "INTEGER", "BYTES", "DIGEST"]
                },
                "namespace": { "type": "string", "example": "testNS" },
                "setName": { "type": "string", "example": "testSet" },
                "userKey": {
                    "type": "object",
                    "example": "userKey",
                    "description": "The user key, it may be a string, integer, or URL safe Base64 encoded bytes."
                }
            },
            "title": "Key"
        },
...
        "Record": {
            "type": "object",
            "properties": {
                "bins": {
                    "type": "object",
                    "example": { "bin1": "val1", "pi": "3.14" },
                    "description": "A mapping from binName to binValue"
                },
                "generation": {
                    "type": "integer",
                    "format": "int32",
                    "example": 2,
                    "description": "The generation of the record."
                },
                "ttl": {
                    "type": "integer",
                    "format": "int32",
                    "example": 1000,
                    "description": "The time to live for the record, in seconds from now."
                }
            },
            "title": "Record"
        },
        "RestClientError": {
            "type": "object",
            "properties": {
                "inDoubt": {
                    "type": "boolean",
                    "example": false,
                    "description": "A boolean specifying whether it was possible that the operation succeeded. This is only included if true."
                },
                "message": {
                    "type": "string",
                    "example": "Error Message",
                    "description": "A message describing the cause of the error."
                }
            },
            "title": "RestClientError",
            "description": "Error object returned from unsuccesful operations."
        },
...
    }
}